axios bug fix
Axios was not sending API requests to the server if refresh token failed to validate. This prevented the client from communicating with the server if they possessed an invalid web token.

removed the return reject line when token failed to validate so client can communicate to server with invalid credentials.

helpful if the client wants to do something like obtain valid credentials through signing in